feat: default owner and repository to Git origin (#341)
## PR Checklist

- [x] Addresses an existing open issue: fixes #196
- [x] That issue was marked as [`status: accepting
prs`](https://github.com/JoshuaKGoldberg/template-typescript-node-package/issues?q=is%3Aopen+is%3Aissue+label%3A%22status%3A+accepting+prs%22)
- [x] Steps in
[CONTRIBUTING.md](https://github.com/JoshuaKGoldberg/template-typescript-node-package/blob/main/.github/CONTRIBUTING.md)
were taken

## Overview

Calls to `git remote -v` to get the repo URL, and parses that as
`--owner` and `--repository` field defaults.

Note that this doesn't add `--description` as a default because I don't
want to mess with calling to the GitHub API & potentially needing to
authenticate. If someone really wants to add that in I'd encourage them
to file a separate feature request issue. 🙂