feat: offer to create or use a different repository in setup (#411)
## PR Checklist

- [x] Addresses an existing open issue: fixes #183
- [x] That issue was marked as [`status: accepting
prs`](https://github.com/JoshuaKGoldberg/template-typescript-node-package/issues?q=is%3Aopen+is%3Aissue+label%3A%22status%3A+accepting+prs%22)
- [x] Steps in
[CONTRIBUTING.md](https://github.com/JoshuaKGoldberg/template-typescript-node-package/blob/main/.github/CONTRIBUTING.md)
were taken

## Overview

Uses a `while (true)` to continuously check if the `repository` exists.
If it doesn't, the user can either:

* Bail out (`"bail"`)
* Create a new repository (`"create"`)
* Try again with a different repository (`"different"`)

Moves the `octokit` creation logic up in the file so that it can be used
for repository checking. This is necessary because the `owner` might be
private organization/user.