feat: added a 'common' config base, with more docs (#849)
## PR Checklist

- [x] Addresses an existing open issue: fixes #835
- [x] That issue was marked as [`status: accepting
prs`](https://github.com/JoshuaKGoldberg/create-typescript-app/issues?q=is%3Aopen+is%3Aissue+label%3A%22status%3A+accepting+prs%22)
- [x] Steps in
[CONTRIBUTING.md](https://github.com/JoshuaKGoldberg/create-typescript-app/blob/main/.github/CONTRIBUTING.md)
were taken

## Overview

Describes the new "common" base as what I imagine most repositories
would want, and gives it the `hint: "recommended"` instead of the
"everything" base. And adds some more test coverage to
`augmentOptionsWithExcludes`.

I think Knip might technically belong in the "everything" base per the
current descriptions... but gosh darn is it so useful. I wonder what's
better: including it in "common" so everyone gets the benefit? Or
pushing it to "everything" so as a reward for trying that one out? For
now I think the former is more likely to be beneficial.

```plaintext
│
◆  How much tooling would you like the template to set up for you?
│  ○ everything  The most comprehensive tooling imaginable: sorting, spellchecking, and more!
│  ● essentials  Bare starters plus testing and automation for all-contributors and releases. (recommended)
│  ○ minimum     Just bare starter tooling: building, formatting, linting, and type checking.
│  ○ prompt      (allow me to customize)
└
```