feat: add --offline mode (#879)
## PR Checklist

- [x] Addresses an existing open issue: fixes #878
- [x] That issue was marked as [`status: accepting
prs`](https://github.com/JoshuaKGoldberg/create-typescript-app/issues?q=is%3Aopen+is%3Aissue+label%3A%22status%3A+accepting+prs%22)
- [x] Steps in
[CONTRIBUTING.md](https://github.com/JoshuaKGoldberg/create-typescript-app/blob/main/.github/CONTRIBUTING.md)
were taken

## Overview

Adds `--offline` flag that skips running scripts. That includes a
timeout for `npm whoami`, as I've found that command to hang when I'm on
an airplane.

Also corrects a bit of the logic around adding the user as a
contributor, while I'm in the area. And fixes up unit tests around that.