chore: use set-github-repository-labels and populate-all-contributors-for-repository internally (#1772)
## PR Checklist

- [x] Addresses an existing open issue: fixes #1768
- [x] That issue was marked as [`status: accepting
prs`](https://github.com/JoshuaKGoldberg/create-typescript-app/issues?q=is%3Aopen+is%3Aissue+label%3A%22status%3A+accepting+prs%22)
- [x] Steps in
[CONTRIBUTING.md](https://github.com/JoshuaKGoldberg/create-typescript-app/blob/main/.github/CONTRIBUTING.md)
were taken

## Overview

Brings on a few utilities I recently extracted:

* https://github.com/JoshuaKGoldberg/octokit-from-auth
*
https://github.com/JoshuaKGoldberg/populate-all-contributors-for-repository
* https://github.com/JoshuaKGoldberg/set-github-repository-labels

These should all have the same functionality as before. This should be a
`chore:` PR, not a `feat:` or `fix:`.

The utility here is that the latter two are now used as standalone
scripts in onboarding to the "create" engine. I thought they'd be done
as network requests (#1752), but in this way I found it a lot simpler to
have them as standalone scripts (#1751,
https://github.com/JoshuaKGoldberg/create/issues/60).

💖