fix: don't include chore PRs in changelog
This change adjusts the release-it config for both this repo, as well as what the create script generates, so that only three categories of changes are included in the changelog entries: feat, fix, and perf.  The idea being that these are the only three types of changes that are generally user-facing.