fix: don't include chore PRs in GH release page
This change adjusts the release-it config for both this repo, as well as what the create script generates, so that the previous changes to enhance the changelog generation flows through to the github release page too.

I discussed this here release-it/release-it#1025 (comment), and @webpro theorized it was the autoGenerate that overrode the changelog output from being applied to the GH release too.

I created a test repo to verify and confirmed that removing `autoGenerate`, indeed, fixed the issue: https://github.com/michaelfaith/mf-cta-testing/releases
2.0.2 is after this change, which included both a chore change and fix change, but changelog and GH release page only show the bug fix (notice also the group heading, which isn't in the previous releases). 2.0.1 is before the change, with the original configuration