feat: bumped to bingo@0.5.8 with serializable requests (#1978)
## PR Checklist

- [x] Addresses an existing open issue: fixes #1977
- [x] That issue was marked as [`status: accepting
prs`](https://github.com/JoshuaKGoldberg/create-typescript-app/issues?q=is%3Aopen+is%3Aissue+label%3A%22status%3A+accepting+prs%22)
- [x] Steps in
[CONTRIBUTING.md](https://github.com/JoshuaKGoldberg/create-typescript-app/blob/main/.github/CONTRIBUTING.md)
were taken

## Overview

As described in
https://github.com/JoshuaKGoldberg/create-typescript-app/issues/1977#issuecomment-2718605049,
adds an `existingLabels` option that reads in the existing labels for a
repository. If that fails or there isn't a repo yet (e.g. in setup mode)
then
[`github-default-labels`](https://github.com/JoshuaKGoldberg/github-default-labels)
supplies the default labels. I would have wanted to use the GitHub API
to fetch the default labels for the owner but that API doesn't seem to
exist.

Anyway, once those existing labels exist, they're passed to
[`determineLabelChange`](https://github.com/JoshuaKGoldberg/set-github-repository-labels/?tab=readme-ov-file#determinelabelchanges)
to get which requests to make. Those are mapped to Bingo's requests
format.

💝