feat: standalone CTA Transitions workflow (#2012)
<!-- 👋 Hi, thanks for sending a PR to create-typescript-app! 🎁.
Please fill out all fields below and make sure each item is true and [x]
checked.
Otherwise we may not be able to review your PR. -->

## PR Checklist

- [x] Addresses an existing open issue: fixes #2009
- [x] That issue was marked as [`status: accepting
prs`](https://github.com/JoshuaKGoldberg/create-typescript-app/issues?q=is%3Aopen+is%3Aissue+label%3A%22status%3A+accepting+prs%22)
- [x] Steps in
[CONTRIBUTING.md](https://github.com/JoshuaKGoldberg/create-typescript-app/blob/main/.github/CONTRIBUTING.md)
were taken

## Overview

Moves the transition mode logic into two files:

1. `.github/actions/transition/action.yml`: A dedicated composite action
for running in transition mode and checking in & commenting any changes
2. `.github/workflows/cta.yml`: Checks if the action should be run, and
runs it if so

This means the job won't be marked as _Skipped_ if the logic is skipped.
It'll still be _Success_ and thus not block PRs.

🎁