fix: support package.json bin objects (#2040)
## PR Checklist

- [x] Addresses an existing open issue: fixes #2039
- [x] That issue was marked as [`status: accepting
prs`](https://github.com/JoshuaKGoldberg/create-typescript-app/issues?q=is%3Aopen+is%3Aissue+label%3A%22status%3A+accepting+prs%22)
- [x] Steps in
[CONTRIBUTING.md](https://github.com/JoshuaKGoldberg/create-typescript-app/blob/main/.github/CONTRIBUTING.md)
were taken

## Overview

Updates the `PartialPackageData` type to note that `bin` could be a
`Record<string, string>`, not just a `string`. `blockPackageJson` will
collect its values if it is an object.

Also updates other references to `options.bin` to use a new
`getPrimaryBin` function for determining what the main runner is. Per
https://docs.npmjs.com/cli/v11/configuring-npm/package-json#bin using
the same name as the repo is equivalent to using a string.

🎁