fix: correct README.md parsing without an existing Usage h2 (#2143)
## PR Checklist

- [x] Addresses an existing open issue: fixes #2140
- [x] That issue was marked as [`status: accepting
prs`](https://github.com/JoshuaKGoldberg/create-typescript-app/issues?q=is%3Aopen+is%3Aissue+label%3A%22status%3A+accepting+prs%22)
- [x] Steps in
[CONTRIBUTING.md](https://github.com/JoshuaKGoldberg/create-typescript-app/blob/main/.github/CONTRIBUTING.md)
were taken

## Overview

https://github.com/JoshuaKGoldberg/prune-github-notifications didn't
have a `## Usage` in its README.md, and instead of allowing all content
until the next relevant h2 or end-of-file, the option was defaulting to
`undefined`.

This reminds me that `blockREADME` hardcodes a `## Usage` in there -
that's not an assumption I want to keep. But removing that assumption
can be a followup.

🎁