fix: correct blockCSpell intake ignorePaths (#2174)
## PR Checklist

- [x] Addresses an existing open issue: fixes #2173
- [x] That issue was marked as [`status: accepting
prs`](https://github.com/JoshuaKGoldberg/create-typescript-app/issues?q=is%3Aopen+is%3Aissue+label%3A%22status%3A+accepting+prs%22)
- [x] Steps in
[CONTRIBUTING.md](https://github.com/JoshuaKGoldberg/create-typescript-app/blob/main/.github/CONTRIBUTING.md)
were taken

## Overview

Switches the Addon name to `ignorePaths` so that everything is called
the same. The alternative was using `ignorePaths` for the schema that
parses in files but `ignores` for the Addons schema.

🎁