feat: switch from deprecated `eslint-plugin-markdown` to `@eslint/markdown` (#2297)
## PR Checklist

- [x] Addresses an existing open issue: fixes #2273
- [x] That issue was marked as [`status: accepting
prs`](https://github.com/JoshuaKGoldberg/create-typescript-app/issues?q=is%3Aopen+is%3Aissue+label%3A%22status%3A+accepting+prs%22)
- [x] Steps in
[CONTRIBUTING.md](https://github.com/JoshuaKGoldberg/create-typescript-app/blob/main/.github/CONTRIBUTING.md)
were taken

## Overview

Swaps in-place from the old plugin to the new.

This isn't _quite_ what the issue suggested, but I plan on sending a
fast follow for the more full #1926 to migrate off of Markdownlint. I
think it'll be nice to have a reference version of CTA that uses
`@eslint/markdown` instead of `eslint-plugin-markdown` _before_ the
removal of Markdownlint. Just in case it's useful.

🎁